[bugfix] remove launch_lb.py (rust_LB has been merged in sgl-router, no need for launch_lb)
