+++ .gitignore
@@ -235,3 +235,6 @@ compile_commands.json
 Cargo.lock
 
 lmms-eval
+
+# lingma
+.lingma
